# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Gym Ledger API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/members/member_detail/{id}": {
      "get": {
        "summary": "Get member details by ID",
        "description": "Fetch a single member's details using their unique ID",
        "tags": [
          "Members"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Member ID",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Member details fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "msg": {
                      "type": "string",
                      "example": "Member fetched successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "example": 1
                        },
                        "name": {
                          "type": "string",
                          "example": "John Doe"
                        },
                        "email": {
                          "type": "string",
                          "example": "john@example.com"
                        },
                        "contact": {
                          "type": "string",
                          "example": "9876543210"
                        },
                        "emergency_contact": {
                          "type": "string",
                          "example": "9123456780"
                        },
                        "height": {
                          "type": "number",
                          "format": "float",
                          "example": 175.5
                        },
                        "weight": {
                          "type": "number",
                          "format": "float",
                          "example": 70.2
                        },
                        "fitness_goal": {
                          "type": "boolean",
                          "example": true
                        },
                        "personal_training": {
                          "type": "boolean",
                          "example": false
                        },
                        "medical_condition": {
                          "type": "string",
                          "example": "None"
                        },
                        "join_date": {
                          "type": "string",
                          "format": "date-time",
                          "example": "2025-03-10T10:00:00.000Z"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "401": {
            "description": "Unauthorized access"
          },
          "404": {
            "description": "Member not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/membership_plane/add": {
      "post": {
        "summary": "Create a new membership plan",
        "tags": [
          "Membership Plan"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "plan_name",
                  "plan_duration",
                  "amount"
                ],
                "properties": {
                  "plan_name": {
                    "type": "string",
                    "example": "Gold Plan"
                  },
                  "plan_duration": {
                    "type": "string",
                    "example": "3 Months"
                  },
                  "amount": {
                    "type": "number",
                    "example": 1500,
                    "description": "Must be greater than 0"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Plan created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "msg": {
                      "type": "string",
                      "example": "Plan created successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error or duplicate plan",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "msg": {
                      "type": "string",
                      "examples": {
                        "required": {
                          "value": "All fields are required"
                        },
                        "zeroAmount": {
                          "value": "Amount can not be 0"
                        },
                        "duplicate": {
                          "value": "Plan already exists"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/membership_plane/delete/{id}": {
      "delete": {
        "summary": "Delete a membership plan",
        "tags": [
          "Membership Plan"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Membership Plan ID to delete",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plan deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "msg": {
                      "type": "string",
                      "example": "Plan deleted successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID or deletion failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string",
                      "examples": {
                        "invalidId": {
                          "value": "Invalid ID"
                        },
                        "notFound": {
                          "value": "Plan not found"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/membership_plane/get": {
      "get": {
        "summary": "Get all membership plans",
        "tags": [
          "Membership Plan"
        ],
        "responses": {
          "200": {
            "description": "Successfully fetched membership plans",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "msg": {
                      "type": "string",
                      "example": "Plans fetched successfully"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "p_id": {
                            "type": "integer",
                            "example": 1
                          },
                          "plan_name": {
                            "type": "string",
                            "example": "Gold Plan"
                          },
                          "plan_duration": {
                            "type": "string",
                            "example": "3 Months"
                          },
                          "amount": {
                            "type": "number",
                            "example": 1500
                          },
                          "created_by_id": {
                            "type": "integer",
                            "example": 2
                          },
                          "created_at": {
                            "type": "string",
                            "format": "date-time",
                            "example": "2026-03-20T10:00:00.000Z"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Failed to fetch plans",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string",
                      "example": "Failed to fetch plans"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  },
...
}